Split ManageManifestModal into multiple files
